+++ package.json
@@ -13,6 +13,7 @@
 		"axios": "^0.24.0",
 		"cheerio": "^1.0.0-rc.10",
 		"cors": "^2.8.5",
-		"express": "^4.17.2"
+		"express": "^4.17.2",
+		"puppeteer": "^13.0.1"
 	}
 }
